Decorate ASTs with hashes of each node.
The fingerprinting module implements the algorithm described in M. Chilowicz,
E. Duris, G. Roussel (2009) Syntax Tree Fingerprinting: A Foundation For Source
Code Similarity Detection. This is used by the GumTree matcher algorithm.

The implementation here is based closely on the GumTree implementation.
Implement to_short_string() and to_static_hash_string() on NodeId.
Compute tree isomorphism based on hashes (as a method on matcher objects).
The hash generator here is based on the default GumTree hash generator,
i.e. the Md5RollingHashGenerator class.

In addition, we create a new enum in ast.rs, FromOrToNodeId, which can be
used by functions and methods which need to match on either a NodeId<FromNodeId>
or a NodeId<ToNodeId> type.