Initial things added to induced graph stage. This does not include tests right now.
Added a buggy mapper. The dot mapper works but there's a slight bug which makes duplicates which needs to be fixed.

Adding the changes. Made follow up would the merge conflict.

Wrap a RefCell around to_arena in MappingStore.
This is currently not necessary in the master branch, but is needed for
changes in a feature branch, which fix a bug in the edit script generator.