Calculate a q-gram distance between strings.
Implementation is based on Java simmetrics library
https://github.com/Simmetrics/simmetrics

Uses char::REPLACEMENT_CHARACTER for padding. This is intended to
reduce the liklihood of collisions, where the padding character
appears in the input string. Padding characters do not appear in
input strings. If the default padding character appears in a string
to be tokenised, search for an alternative. If no alternative can be
found in a reasonable time, panic.