Carefully compare f64 values for equality.
Since floating point numbers are not exact, in the general case we do not
wish to use a simple comparison. In general, we use a relative error, which
compares the difference between the two arguments to their magnitude. This
captures the case where the arguments are very large, but their difference
might be bigger than machine epsilon. However, when the inputs are close to
zero, a relative error can result in a division by zero or a NaN, and so in
this case we use an absolute difference. Note that this function will return
`false` if the inputs are both very small but differently signed, even if
they are both the smallest possible positive / negative f64s.

We use the new `f64_eq` function wherever we test floating point numbers
for equality. Note that one value in the test cases for the `qgram`
module has changed. The old value was erroneous (the test data was
originally generated as a Java float, not a double).